refine mod test_io_error
