+++ Cargo.toml
@@ -52,7 +52,7 @@ protobuf = { git = "https://github.com/pingcap/rust-protobuf", rev = "82b49fea7e
 protobuf-codegen = { git = "https://github.com/pingcap/rust-protobuf", rev = "82b49fea7e696fd647b5aca0a6c6ec944eab3189" }
 
 [workspace]
-members = [ "stress" ]
+members = ["stress"]
 
 [[bench]]
 name = "bench_recovery"
@@ -60,7 +60,6 @@ harness = false
 required-features = ["failpoints"]
 
 [[test]]
-name = "io_error_test"
-path = "failpoints/io_error_test.rs"
-harness = true
+name = "failpoints"
+path = "failpoints/mod.rs"
 required-features = ["failpoints"]
