+++ failpoints/mod.rs
@@ -1,0 +1,1 @@
+mod test_io_error;
